Add files via upload
Project goal is to predict salary of the MBA graduate based on academic details. At high level jobs, it is very important for the companies to recruit best talent and offer a good compensation. So, project provides candidate analysis to hiring personnel for recognition of talent and their market value.